pyproject.toml
[project]
name = "accountability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastmcp>=2.14.2",
    "langchain>=1.2.3",
    "langchain-huggingface>=1.2.0",
    "psycopg2-binary>=2.9.11",
]
